#6 업데이트 고정 및 행아웃 참여 기능 추가, 앱 아이콘 업데이트

+++ lib/utils/image_picker_utils.dart
@@ -22,8 +22,8 @@ class ImagePickerUtils {
 
       return await _picker.pickImage(
         source: source,
-        maxWidth: maxWidth,
-        maxHeight: maxHeight,
+        maxWidth: maxWidth.toDouble(),
+        maxHeight: maxHeight.toDouble(),
         imageQuality: imageQuality,
       );
     } catch (e) {
@@ -55,8 +55,8 @@ class ImagePickerUtils {
       try {
         images = await _picker.pickMultiImage(
           imageQuality: imageQuality,
-          maxWidth: maxWidth,
-          maxHeight: maxHeight,
+          maxWidth: maxWidth.toDouble(),
+          maxHeight: maxHeight.toDouble(),
           limit: limit,
         );
       } catch (platformException) {
